Swap locations of CDP and Lending Pool in UI
